forum_backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
bleach==6.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.1
daphne==4.1.2
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-celery-beat==2.7.0
django-channels==0.7.0
django-cors-headers==4.5.0
django-oauth-toolkit==3.0.1
django-rest-auth==0.9.5
django-timezone-field==7.0
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jwcrypto==1.5.6
kombu==5.4.2
msgpack==1.1.0
oauthlib==3.2.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==5.1.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
tinycss2==1.3.0
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zope.interface==7.1.0
